Datenbank Klasse in eigene Datei verschoben. Klassen gehören normal in eine eigene Datei. Vorbereitung für den Blocklist Patch. Hier wird nur die DB Klasse benötigt und nicht der Rest der Funktionen.
