+++ function.php
@@ -1,5 +1,8 @@
 <?php
 
+/* Include Database class */
+include "db.class.php";
+
 const hood_mysql_fields = '
 	ID as id,
 	name,
@@ -101,36 +104,6 @@ function debug($msg)
 	if (DEBUG) {
 		print_r($msg);
 		echo "\n";
-	}
-}
-
-/**
- * Singelton DB instance
- */
-class db
-{
-	private static $instance = NULL;
-
-	private function __construct()
-	{
-	}
-
-	public static function getInstance()
-	{
-		if (!self::$instance) {
-			try { // catch errors and do not show exception details
-				require('config.inc.php');
-				self::$instance = new PDO('mysql:host=' . $mysql_server . ';dbname=' . $mysql_db, $mysql_user, $mysql_pass);
-				self::$instance->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
-			} catch (PDOException $e) {
-				exit(showError(500, "Error while connecting to database."));
-			}
-		}
-		return self::$instance;
-	}
-
-	private function __clone()
-	{
 	}
 }
